Hi
This change enables the CDCACM to send with around 1M.
With -Os the buffer send should be at lest around 32byte-
With -O3 already with 4 bytes 1M is reached.
It is only reached when there is only one USB device registers in the USBManager.
It depents on PR #446