@EmbeddedMan 2019-03-12
ok to test

On Tue, Mar 12, 2019 at 4:26 PM chipKIT Bot <notifications@github.com>
wrote:

> Can one of the admins verify this patch?
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/chipKIT32/chipKIT-core/pull/447#issuecomment-472187657>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAbeCEkfTY75oeClmTTGqvvnC12UPjkqks5vWBtrgaJpZM4br6Py>
> .
>