@majenkotech 2019-03-13
This looks good to me. About 850k/s with two CDCACMs loaded going to Minicom.

Vote to merge.