@majenkotech 2019-03-13
Hold up - on the Lenny I'm seeing some data loss.  Using Paul S's USB test code I get about 550k/s on a WiFire. But on the Lenny I see this:

```
USB Fast Serial Transmit Bandwidth Test, capture this text.
USB Fast Serial Transmit Bandwidth Test, capture this text.
USB Fast Serial Transmit Bandwidth Test, capture this text.
USB Fas Serial Transmit Bandwidth Test, capture this text.
USB Fast Serial Transmit Bandwidth Test, capture this text.
USB Fast Serial Transmit Bandwidth Test, capture this text.
USB Fast Serial Transmit Bandwidth Test, capture this text.
USB Fast Serial Transmit Bandwidth Test, capture this text.
USB Fast Serial Transmit Bandwidth Test, capture this text.
```

Every so often there's a letter missing from the data.

**Hold on merge for now**