@EmbeddedMan 2019-03-13
OK, good to know. We'll hold until the data loss issue is solved.