@majenkotech 2019-03-13
The problem possibly stems from #446 as I am getting random "1" inserted with this test code on that PR.