@majenkotech 2019-03-14
Yep, that's fixed it.  I think we're good to go with this PR now.